Move TimeSpanTests
